move const variables and helpers to their scope
