+++ imjoy/env.py
@@ -6,7 +6,7 @@ import subprocess
 import sys
 import uuid
 
-from imjoy.helper import kill_process
+from imjoy.util import kill_process
 
 
 def prepare_env(opt, logger):
