+++ imjoy/services/terminal.py
@@ -8,8 +8,9 @@ import subprocess
 import sys
 
 from imjoy.connection.decorator import socketio_handler as sio_on
-from imjoy.const import NAME_SPACE, __version__
-from imjoy.helper import get_psutil
+from imjoy.connection.server import NAME_SPACE
+from imjoy.engine import __version__
+from imjoy.util import get_psutil
 
 if sys.platform != "win32":
     import fcntl
